Re-factored update check
