+++ _caster.py
@@ -10,11 +10,15 @@ from dragonfly.windows.window import Window
 
 logging.basicConfig(format = "%(asctime)s : %(levelname)s : %(funcName)s\n%(msg)s")   
 
-from castervoice.lib import settings  # requires toml
-settings.initialize()
-
 from castervoice.lib.ctrl.dependencies import DependencyMan  # requires nothing
 DependencyMan().initialize()
+
+from castervoice.lib import settings  
+settings.initialize()
+
+from castervoice.lib.ctrl.updatecheck import UpdateChecker # requires settings/dependencies
+UpdateChecker().initialize()
+
 _NEXUS = None
 
 class LoggingHandler(logging.Handler):
